Update tests for posts page
